pyproject.toml
[project]
name = "specify-cli"
version = "0.0.22"
description = "Specify CLI, part of GitHub Spec Kit. A tool to bootstrap your projects for Spec-Driven Development (SDD)."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Specify CLI Contributors"},
]
keywords = [
    "spec-driven-development",
    "sdd",
    "specification",
    "development",
    "azure",
    "bicep",
    "infrastructure-as-code",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx[socks]>=0.25.0",
    "platformdirs>=3.0.0", 
    "readchar>=4.0.0",
    "truststore>=0.10.4",
]

[project.optional-dependencies]
# Azure Bicep Generator feature dependencies
bicep = [
    "azure-identity>=1.15.0",
    "azure-mgmt-resource>=22.0.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
]

# Azure Bicep Validation feature dependencies
validation = [
    "httpx[socks]>=0.25.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
    "pytest-asyncio>=0.21.0",
    "scikit-learn>=1.3.0",  # For learnings database semantic similarity (TF-IDF)
]

# Development dependencies
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

# Testing dependencies
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.2.0",
    "pytest-xdist>=3.5.0",
    "pytest-httpx>=0.21.0",
    "pytest-mock>=3.12.0",
    "coverage[toml]>=7.3.0",
]

# All optional dependencies
all = [
    "specify-cli[bicep,validation,dev,test]",
]

[project.scripts]
specify = "specify_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specify_cli"]

